requirements.txt
asgiref==3.8.1
Django==5.2.3
django-phonenumber-field==8.1.0
django-unfold==0.60.0
phonenumbers==9.0.8
pillow==11.2.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
